# src/main/java/com/finalproject/backend/openapi/absences.yml
openapi: 3.0.3
info:
  title: Absences API
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /api/absences:
    post:
      summary: Create absence
      tags: [Absences]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AbsenceCreateDTO'
            example:
              studentId: 12
              classCourseId: 7
              date: "2026-01-04"
      responses:
        '200':
          description: Absence created

  /api/absences/{id}:
    get:
      summary: Get absence by id
      tags: [Absences]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          example: 5
      responses:
        '200':
          description: Absence found
        '404':
          description: Absence not found

  /api/absences/student/{studentId}:
    get:
      summary: Get all absences of a student (grouped by course)
      tags: [ Absences ]
      security:
        - bearerAuth: [ ]
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Student absences
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /api/absences/student/{studentId}/course/{classCourseId}:
     get:
      summary: Get absences of a student for a specific course
      tags: [ Absences ]
      security:
         - bearerAuth: [ ]
      parameters:
         - name: studentId
           in: path
           required: true
           schema:
            type: integer
           example: 1
         - name: classCourseId
           in: path
           required: true
           schema:
            type: integer
           example: 7
      responses:
          '200':
            description: Student absences for course
          '401':
            description: Unauthorized
          '403':
            description: Forbidden

  /api/absences/course/{classCourseId}:
    get:
      summary: Get all absences for a course (all students)
      tags: [ Absences ]
      security:
        - bearerAuth: [ ]
      parameters:
        - name: classCourseId
          in: path
          required: true
          schema:
            type: integer
          example: 7
      responses:
          '200':
            description: Course absences
          '401':
            description: Unauthorized
          '403':
            description: Forbidden

  /api/absences/classroom/{classroomId}:
    get:
      summary: Get all absences for a classroom
      tags: [ Absences ]
      security:
        - bearerAuth: [ ]
      parameters:
        - name: classroomId
          in: path
          required: true
          schema:
            type: integer
          example: 3
      responses:
          '200':
            description: Classroom absences
          '401':
            description: Unauthorized
          '403':
            description: Forbidden

    put:
      summary: Update absence
      tags: [Absences]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          example: 5
      requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AbsenceUpdateDTO'
              example:
                date: "2026-01-04"
                excused: true
      responses:
        '200':
          description: Absence updated

    delete:
      summary: Delete absence
      tags: [Absences]
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          example: 5
      responses:
        '204':
          description: Absence deleted

components:
  schemas:
    AbsenceCreateDTO:
      type: object
      required: [studentId, classCourseId, date]
      properties:
        studentId:
          type: integer
          example: 12
        classCourseId:
          type: integer
          example: 7
        date:
          type: string
          format: date
          example: "2026-01-04"

    AbsenceUpdateDTO:
      type: object
      required: [ date, excused ]
      properties:
        date:
          type: string
          format: date
          example: "2026-01-04"
        excused:
          type: boolean
          example: true
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT